Refactor out AsyncHelper.RunSync() calls to assist diagnosis of app hangs
As per posts by @jmhinnen on 20 Feb 2019, it's possible that some app hangs are due to sync-over-async calls locking up.
Microsoft support suggested refactoring to ensure call stacks are all async or sync, and not a mix of the two.
This commit removes all AsyncHelper.RunSync() calls where possible to help identify if this is indeed the source of the problem.

Notable exceptions where AsyncHelper.RunSync() can't trivially be removed:
 - AsyncBackgroundJob
 - NotificationDistributionJob.Execute() -> SignalR async distribution
 - SignalRRealTimeNotifier.SendNotifications() -> signalRClient.SendAsync()
 - EventBus.Trigger() on async events
 - AbpLogInManagerExtensions.Login() -> relies on lots of async calls to check password, lockout, etc.

References: #4303